Implement figure CRUD

+++ beetle/dao/figureDao.js
@@ -15,21 +15,21 @@ class FigureDao {
     }
 
     findAll(userId) {
-        let sqlRequest = "SELECT * FROM figure WHERE user_id=$user_id"
+        let sqlRequest = "SELECT id FROM figure WHERE user_id=$user_id"
         let sqlParams = {$user_id: userId}
 
         return this.common.findAll(sqlRequest, sqlParams).then(rows => {
             let figures = [];
             for (const row of rows) {
-                figures.push(new Figure(row.id, row.code, row.user_id))
+                figures.push(row.id)
             }
             return figures
         })
     }
 
     update(Figure) {
-        let sqlRequest = "UPDATE figure SET code=$code WHERE id=$id"
-        let sqlParams = {$id: Figure.id, $code: Figure.code}
+        let sqlRequest = "UPDATE figure SET code=$code WHERE id=$id and user_id=$user_id"
+        let sqlParams = {$id: Figure.id, $code: Figure.code, $user_id: Figure.userId}
 
         return this.common.run(sqlRequest, sqlParams)
     }
@@ -41,9 +41,9 @@ class FigureDao {
         return this.common.run(sqlRequest, sqlParams)
     }
 
-    deleteById(id) {
-        let sqlRequest = "DELETE FROM figure WHERE id=$id"
-        let sqlParams = {$id: id}
+    deleteById(id, userId) {
+        let sqlRequest = "DELETE FROM figure WHERE id=$id and user_id=$user_id"
+        let sqlParams = {$id: id, $user_id: userId}
         return this.common.run(sqlRequest, sqlParams)
     }
 }
